CEHCIG-436 CEHCIG-437 Split publication state from public view permission.
To be publicly viewable a metadata document needs to be 'published' and view permission given to 'public'.

PermissionService simplified, does not need to check for author because they will be added to the permissions object. Hands off to the MetadataInfo for access decisions.
MetadataInfo has a constructor that allows a copy to be made (used in  the publication workflow state transitions)
MetadataInfo has methods for deciding access permission queries, the permissions are stored on the MetadataInfo so it is the place to decide access queries. Rather than give out data for another object to make the decisions.
MetadataInfo converted permissions to a HashMultiMap so that duplicate permissions cannot be stored, and made sure permissions object is always created in constructor.
MetadataListingService modified to use MetadataInfo to check permissions. It does not need the permission service because it already has the information in the MetadataInfo object.
ViewIndexer deleted as view indexing simplified and only need to check view permissions, not have to check author permissions separately.
TerraCatalog importer adds view permission for the author.