MetadataDocument marshalling now done in one place
Setup of a MetadataDocument when it is read from Git all
now happens in a consistent way in the BundledReaderService.

The controllers and services that need a MetadataDocument can now get
one from the DocumentRepository.

The GitRepoWrapper isolates interaction with Git repository mainly to
ease testing of the DocumentRepository.

CEHCIG-669