unreliable rspec search page test removed
The test to check if content is different between page 2 and page one is
unreliable, it relied on checking that all the content was different
between the pages.

I have removed this test until I can think of a more reliable way of
achieving the desired test outcome.

This test broke the continuous integration

CEHCIG-762-CEH-Model-Application-keywords-multiplying