Add DummyLinkedElterDocument DocumentType to Java
We want to keep the eLTER catalogue to the one DocumentType,
ElterDocument, but also want multiple options in the create dropdown,
including the option to create a linked metadata document.

Rather than rigging up bespoke adjustments to the HTML framework, we add
a placeholder DocumentType which exists to accept a URL and other
parameters, such as a record type. Once submitted, these will be used to
retrieve the metadata from elsewhere and save it as a regular
ElterDocument.

This commit only deals with the Java side - an editor is coming later.