Add linked editor and correct media type bug
An ElterLinkedEditorView is created to allow inputting the minimal
information required for the linked document retrieval functionality. We
leave the linkedDocumentType as free-form until specific options are
decided.

We also add handlers for the dummy documents, which simply save them for
now - we will add the document retrieval service later.

codelist.properties has the linked elter records added. The new
ElterLinkedEditorView is added to Main.coffee.

ConvertUsing and Template annotations are added to the dummy document
type. They were not initially included since this type should never need
to be rendered. However, the message converters in Webconfig call
getSupportedMediaTypes, which requires annotations to avoid spurious
errors.