Add and wire document retrieval service
The LinkedDocumentRetrievalService is added which supplies a get method
to retrieve a json document from a URL and save it as an ElterDocument.
This has been successfully tested with the following EIDC document:
https://catalogue.ceh.ac.uk/documents/fc9bcd1c-e3fc-4c5a-b569-2fe62d40f2f5?format=json

We replace the dummy method in DocumentController, which saves
DummyLinkedElterDocuments as normal, with a call to the new method, the
output of which is then saved as normal.