Pages can use different javascript files
It was becoming very slow developing the Hubbub upload page as page refreshes where downloading many javascript files for all the other parts of the catalogue that were not used by the Upload application.

Now each application can specify the scripts needed.
Need to configure requirejs for the page.

EDS-2436