Refactor linked documents and DocumentController
NOTE: this commit is formed of the feedback to merge request !364, hence
its large scope.

We rename DummyLinkedElterDocument to LinkedElterDocument to conform
with naming conventions of the codebase, as well as the linkedDocument
field of ElterDocument. Additionally, the class has been rewritten as a
Lombok Value with Jackson JsonCreator and JsonProperties.

We remove the linked documents from HashMapDocumentTypeLookupService as
documents will never need to be displayed in the form
?format=linked-document. For similar reasons we also remove the
Object2TemplatedMessageConverter and .mediaType lines from WebConfig.

The LinkedDocumentRetrievalService has an "elter" profile annotation
added and its pointless error catching code removed.

We move the elter content from the master DocumentController to
its own ElterDocumentController, adding an "elter" profile annotation
and removing the pointless edit code for LinkedElterDocuments since
these aren't saved. Since this restores the prior constructor of
DocumentController, we also revert the DocumentControllerTest back to
its previous form.

Finally, we correctly set the linkedDocument, linkedDocumentUri, and
linkedDocumentType fields when saving a linked document, as these were
previously blank.