Update ElterLinkedEditorView
NOTE: this commit completes the feedback to merge request !364.

We remove unnecessary imports for tidiness' sake - possibly performance
too.

Since the title and description fields will be taken from the retrieved
document, we remove the ability to input them here. We leave the title
read-only with a placeholder to provide visual feedback when a document
is successfully retrieved though.

Finally, we make small adjustments to the help texts to reflect the
current status of the two active fields, linkedDocumentType and
linkedDocumentUri.