Add tests for Controller and Service
We add tests for the two new components ElterDocumentController and
LinkedDocumentRetrievalService. These tests pass on my machine as
committed.

For the ElterDocumentController we test creating and editing
ElterDocuments the usual way, as well as creating one through the linked
document functionality.

For the LinkedDocumentRetrievalServiceTest we test its only
functionality, fetching and saving a JSON document as an ElterDocument.
A real EIDC record in JSON format is saved as a resource to be used by
the mock rest server, which can actually be found at the tested uri.

Finally, fix erroneous whitespace in ElterDocumentController.