Add isLinkedDoc to elter.ftlh
Commit b4baebf694 introduced isLinkedDoc in _title.ftlh for rendering
Gemini documents, presumably not realising that elter and ukems
documents also rely on it. As a result, trying to open elter and ukems
documents in the application has consistently resulted in 500 errors.

We fix only the elter documents here by assigning this property as
always false, as another PR deals with ukems. We don't need this
property as elter has its own concept of linked documents.