Update simple UploadController document type
UploadController relied on uploading to GeminiDocuments to get
attributes etc. to render the upload page on a GET request. This fails
when working with elter documents - presumably others as well.

We convert this to work with the base class, AbstractMetadataDocument,
instead for greater flexibility. This has been tested successfully with
elter documents.

Fixes: EDS-3141