Add import interface and draft SITESImportService
With the SITES record import in mind, we add an interface
CatalogueImportService which can be scheduled to create a recurring
import. This relies on several implicit assumptions which are documented
elsewhere and will eventually be added here on merge to develop.

The SITESImportService partially implements the CatalogueImportService
interface. In its current state it just retrieves the sitemap, parses
the list of records and prints out the first five.