Draft getFullRemoteRecord in SITESImportService
runImport() now, for the first three records in the sitemap:
	- fetches them (getFullRemoteRecord)
	- parses them to JSON (getFullRemoteRecord)
	- logs some of their properties (runImport)

Because some records for the SITES import shouldn't be imported, there's
no sense converting a parsed record to an ElterDocument until we've
checked whether it's needed or not. Otherwise we could return an
ElterDocument directly from getFullRemoteRecord.

We also update some logging to info (so it displays by default) and
update FEED ME A STRAY CAT to a nonplaceholder value (oops).