Draft createRecord in SITESImportService
runImport() now, for the first three records in the sitemap:
        - fetches them (getFullRemoteRecord)
        - parses them to JSON (getFullRemoteRecord)
	- saves them in the datastore if they're datasets (NEW)

For now we only save minimal information in the new documents: a
complete mapping of fields will follow later.

In anticipation of future import logic requiring it, we add new fields
to ElterDocuments, namely an "importID" (an arbitrary String to match
remote records with local ones) and importLastModified (metadata to
potentially see if we need to update our local record).

We also restructure the exception handling so that runImport() will
never raise an exception. This seems sensible as errors during an import
shouldn't have any effect on the running of the catalogue as a whole.
This includes more or less finalising the CatalogueImportService
interface and removing draft comments.