Draft updating logic in SITESImportService
runImport() now, for the first three records in the sitemap:
	- fetches them (getFullRemoteRecord)
	- parses them to JSON (getFullRemoteRecord)
	- if they're of type dataset:
	    - saves them in the datastore if they're new (createRecord)
	    - updates the local record if it exists (NEW)

To check which records exist in the local catalogue which have already
been imported from the SITES database, we have to index the importID
field in Solr as a string.

We also rename recordID to localRecordID/remoteRecordID where
appropriate to avoid ambiguity.