Add basic DOI retrieval from Datacite
When creating a "Linked document" through the catalogue interface, if
the input URL (needs renaming) contains a DOI then that DOI is fetched
from the Datacite API instead fetching the input URL.

The record is populated from that with a title and set as a signpost
record to a Level 0 dataset. There's no error handling yet so this will
fail if there is no title.