Finish DOI retrieval from Datacite
We populate the remaining fields from datacite to form a nice baseline.
Some potential future fields/features like DEIMS sites can be added but
they might require some policy on the other end so I think this set
should do for a good while.

We also set a fallback title in case it's missing, which should avoid
any errors. Other fields that don't do checks seem to be guaranteed to
be present on the datacite end.