Add first draft field mapping
Some fields are more complicated, like DEIMS sites being mapped from the
creator field, but here we add a good collection of initial field
mappings.

The duplicated code in both methods was going to be refactored into the
constructor of ElterDocuments but it wasn't working and I don't have
time to figure out why, so I'll come back to that.