Add first test
We add the first of three tests of the SITESImportService for creating a
new document. Relevant setup will only be refactored out once the other
tests are written.

We also move the SITESImportService to the elter package, which was
supposed to happen in commit 6b053934d87a644e9034ff1fff1038ce08cf4776.