Convert CRLF to LF, remove BOM and .gitattributes
To verify this commit, the `--ignore-cr-at-eol` option to the diff/show
commands is useful.

We remove the .gitattributes because it wasn't doing anything. There are
no more .sh or .pem files and text=lf has no effect in current versions
of git - at time of writing I'm running git version 2.39.2.

A byte order mark was present in gmx.xsd, which we also remove since
we're normalising everything to Linux environments which default to
UTF-8 without BOM.