Map DEIMS sites in SITESImportService
To map the SITES station to a DEIMS site we hardcode the mapping of the
"SITES ids" to DEIMS ids, which are retrieved from Solr to account for
possible changes of name - note that the DEIMS id will never change.

Since querying Solr adds some complexity we move the JSON >
ElterDocument logic back into the SITESImportService.