Handle list of identifiers in SITESImportService
It turned out that 2 (at time of writing) records in the source have a
list of identifiers instead of just a string, which led to constant
reimporting as the importId would end up empty. We handle this case in
createDocumentFromJson now.

Since the identifier now has more complex logic we can't easily test
against it in the main import, so we create the ElterDocument
immediately (instead of delegating to createRecord or updateRecord) so
we can get the identifier with getImportId().

While changing updateRecord we also get the remoteRecordId from the
ElterDocument itself. This has the benefit of making the signatures of
both createRecord and updateRecord the same.

Finally we add a test to test this new case and strip some trailing
whitespace from comments in SITESImportServiceTest.