Finish first draft B2shareImportService
As it stands the B2SHARE import will import the first 5 (not all, any
more) records in the "LTER" community. It gets the record DOIs from the
B2SHARE API and imports them from datacite. Most of the new methods are
just tweaked copies of those from SITESImportService.

Since we are reusing the same JSON processing code as the
LinkedDocumentRetrievalService, this is refactored into a method of the
ElterDocument, `importDataciteJson`.

Also add `b2share.api` to application.properties, which was supposed to
be added before.

Jira: ELTER-20