Refactor to use B2SHARE API instead of Datacite
Datacite metadata is incomplete compared to B2SHARE data and B2SHARE
record listings contain the complete records anyway, so we just use the
B2SHARE records directly.

We also implement a blacklist of DOIs which have already been imported
separately to the production catalogue already - these records are
skipped.

The new createDocumentFromJson() method handles the parsing of the
B2SHARE record JSON. The exception is the importId, since we need to
check the DOI of a record before importing it in runImport().

Jira: ELTER-20