Add extra B2SHARE API field mappings
We add some extra fields that either weren't available in Datacite or
just hadn't been mapped yet. These are:
	- created + publication_date
	- open_access
	- proper link as supplemental
	- community_specific...metadata_url (deims site)
	- disciplines + keywords

Of note is the extra code to query for deims sites and to parse all the
different timestamp formats found in the wild at time of developing. We
also change the main link to a direct b2share url instead of going
through doi.org (which matches the other imports) and move the doi.org
link to the supplemental section.

Tweak some declarations of Lists to
	List<foo> bar = new ArrayList<>();
format.

Add "Unknown" as an organisationName to the author so that the author
displays. At time of writing no authors have an institute in the API.

Jira: ELTER-20