Fix URL handling in B2shareImportService
While deployed on staging I noticed an NPE that triggered on the second
run of the import. Checking it made me realise that the initial URL to
call was permanently overwritten during an import run, meaning the while
loop which is always supposed to run at least once was failing.

We make the initial URL fixed and reference a copy on each run, with
more verbose variable names, to address this issue. This fixes the
broken test in the parent commit.

Jira: ELTER-32