Add concepts to GEMET vocabulary
In addition to the themes we also need the concepts. To preserve the
general design of the HttpKeywordVocabulary we allow passing a single
endpoint or a list and update the retrieve method to iterate over them.

For now the vocab is always cleared before any terms are retrieved,
meaning it may be empty if there are errors.

Jira: EMC-19