Add .editorconfig
After a lot of experimenting it seems that it's impossible to set
certain directories to be "preserved" for e.g. test resources. Therefore
we just set global defaults and some filetype exceptions for now.

Note that this commit does not apply the rules to any files - that will
come later. See trailing whitespace in README.md or 2 space indents in
docs/new-catalogue.md as examples.