Remove 'relationships' from returned GeminiDocument
The json produced by the server for a GeminiDocument has RelatedRecords merged into the Relationships array. This is in an overwritten method of AbstractMetadataDocument.

Jena indexing uses Relationships to add related links to the triple store.

Deleting Realtionships from the EditorMetadata model ensures that only the correct links in RelatedRecords are indexed