Reindent to 4 spaces
Indentation was either fixed automatically by rescaling the existing
indentation or, where this wasn't possible, by reindenting the whole
file with vim 9.0. This will mean some stylistic differences remain
between files but that is beyond the scope of this project.

We also convert some stray tabs to spaces in WmsFeatureInfo.java.