search: Fix escaping of facet terms
The previous implementation of bounding box search caused issues when
selecting facets containing special charaacters such as '&'.  Instead
of turning the query parameters into a hash fragment, we fix up the
Backbone logic to properly intercept clicks on the links and handle
the searches in JavaScript rather than round-tripping to the backend.

This also includes a bit of JS modernisation, as well as fixing up
some CoffeeScript-style interpolation that didn't get translated
properly for some reason.

EMC-113