Use UTF-8 encoding in FusekiExportService
Non-ASCII characters in metadata records were being replaced by U+FFFD
REPLACEMENT CHARACTER.  This is because the default encoding for
non-ASCII characters is ISO-8859-1, rather than the modern standard of
UTF-8.  At some point in sending the request, it gets reinterpreted as
(invalid) UTF-8 and the ISO-8859-1 characters are replaced.

We enforce the encoding as UTF-8 by changing the Content-Type to
include the correct charset.

EMC-148