ServiceAgreementPublicationService: Properly handle UKCEH depositors
Internal UKCEH users (more precisely, those with @ceh.ac.uk email
addresses) were unable to submit draft SAs as they did not have the
correct roles set up.  We change the logic in
ServiceAgreementPublicationService::getPublishingRoles to consider the
user a depositor when the catalogue username is equal to either the
depositor contact details (for an external user where their username
is the whole email) or the catalogue username with @ceh.ac.uk appended
(for internal UKCEH users).

DCDEV-77