Make schema.org output group organisations for responsible parties
The authors and points of contact attached to a record might have an
organisation they are affiliated with referenced.  If that
organisation also has an ROR identifier, there's no need to repeat its
definition for each author in the schema.org output, just to ensure
it's included at the end.

EMC-399